refactor: 필요없는 주석 제거, ReviewDetail로 이름 수정
