+++ src/pages/CreatedReviewManagePage/components/index.ts
@@ -2,8 +2,7 @@ export { default as Tabs } from './Tabs'
 export { default as AllResponseReviewByResponser } from './AllResponseByResponser'
 export { default as AllResponseReviewByReceiver } from './AllResponseByReceiver'
 export { default as SortDropDown } from './SortDropDown'
-export { default as ReviewDetailAccordion } from './ReviewDetailAccordion'
+export { default as ReviewDetailAccordion } from './ReviewDetail'
 export { default as ProfileGroup } from './ProfileGroup'
 export { default as QuestionGroup } from './QuestionGroup'
 export { default as NotFoundSearchUser } from './NotFoundSearchUser'
-export { default as SendReviewDetailAccordion } from './SendReviewDetailAccordion'
